+++ app/controllers/meal_plans_controller.rb
@@ -27,10 +27,10 @@ class MealPlansController < ApplicationController
 
   def toggle_cooked
     if @meal_plan.cooked?
-      restore_inventory(@meal_plan)
+      @meal_plan.restore_inventory!
       @meal_plan.update!(cooked: false)
     else
-      deduct_inventory(@meal_plan)
+      @meal_plan.deduct_inventory!
       @meal_plan.update!(cooked: true)
     end
     respond_with_cell(@meal_plan.reload)
@@ -95,28 +95,6 @@ class MealPlansController < ApplicationController
         )
       end
       format.html { redirect_to kitchen_path }
-    end
-  end
-
-  def deduct_inventory(meal_plan)
-    if meal_plan.recipe
-      meal_plan.recipe.recipe_ingredients.includes(food_item: :pantry_item).each do |ri|
-        ri.food_item.pantry_item&.decrement!(ri.quantity * ri.food_item.servings_per_unit)
-      end
-    end
-    meal_plan.meal_plan_items.includes(food_item: :pantry_item).each do |item|
-      item.food_item.pantry_item&.decrement!(item.quantity * item.food_item.servings_per_unit)
-    end
-  end
-
-  def restore_inventory(meal_plan)
-    if meal_plan.recipe
-      meal_plan.recipe.recipe_ingredients.includes(food_item: :pantry_item).each do |ri|
-        ri.food_item.pantry_item&.increment!(ri.quantity * ri.food_item.servings_per_unit)
-      end
-    end
-    meal_plan.meal_plan_items.includes(food_item: :pantry_item).each do |item|
-      item.food_item.pantry_item&.increment!(item.quantity * item.food_item.servings_per_unit)
     end
   end